--- interfaces/google/google_oauth2.go
@@ -1,7 +1,0 @@
-package google
-
-import "golang.org/x/oauth2"
-
-type GoogleOauth2 interface {
-	GetConnect() *oauth2.Config
-}
